+++ src/input_handler.cpp
@@ -25,6 +25,7 @@ void InputHandler::initializeDefaultBindings() {
     
     // Actions
     keyBindings["."] = InputAction::WAIT;
+    keyBindings["o"] = InputAction::OPEN_DOOR;
     keyBindings["Return"] = InputAction::CONFIRM;
     keyBindings["Escape"] = InputAction::CANCEL;
     keyBindings["q"] = InputAction::QUIT;